review and refine. Test MdUpdt, LossPrint and AccuracyPrint

+++ oneflow/core/kernel/normal_model_update_kernel.cuh
@@ -4,7 +4,7 @@
 namespace oneflow {
 template<typename T>
 __host__ __device__ T RegularizeDiff(T diff, T total_instance_num, T l1, T l2, T pre_model_val) {
-  return diff / static_cast<int64_t>(total_instance_num)
+  return diff / static_cast<int32_t>(total_instance_num)
          + l1 * ((pre_model_val >= 0) - (pre_model_val <= 0)) + l2 * pre_model_val;
 }
 }  // namespace oneflow
